Rewrote serialize and deserialize to be recursive
Instead of working exclusively on the format for the config file, I've
decided to just make it operate recursively. With a growing set of
features, ensuring that the correct cases for the config file are always
handled properly by the serialization functions is more effort than it's
worth, so instead they operate recursively. Any arbitrary EDN structure
should work on them, and the symbols with reserved characters will get
changed to strings.